Refactor 2D/3D conversion
Switch from 'placeIn3d' and 'projectInto2d' (defined in 2D and 3D
modules respectively for a given type) to 'placeOnto' and 'projectInto'
functions both defined in the 3D module. This means that the 2D module
does not have to 'know' anything about 3D, which makes module
dependencies simpler, and means that the functions can be right beside
each other in the documentation.