Switch 'plus' -> 'sum', 'minus' -> 'difference'
These functions are more likely to be used fully applied than
piped/chained, so 'sum' and 'difference' will read better most of the
time. The 'normal' argument order is also a bit easier to explain, and
is consistent with 'dotProduct' and 'crossProduct' (the only other
remaining functions in the 'Arithmetic' section).