Add a couple more candidate tests
These tests are mutually exclusive, but right now they *both* fail and
ideally we'd find a way to make one of them pass. I'm actually leaning
towards "intersectionDoesNotFindCoincidentEndpoints", so that the
behaviour of LineSegment2d.intersection can be described simply as "find
the intersection point if the two segments actually cross each other"
(not just if they touch each other).