Rename Types module to Internal
This will help with updating other packages that are importing the Types module currently. It also will make it possible to re-add a Types module in the future that simply aliases all the types in the Internal module; this may be necessary depending on what rules Elm adopts in the future about how types must be exposed from packages.