Add Direction2d.rotate[Clockwise,Counterclockwise]
`rotateClockwise` is convenient way to get an X direction from a Y direction (current alternatives include `Direction2d.rotateBy (degrees -90)` which involves sin/cos or `Direction2d.perpendicularTo >> Direction2d.flip` which is awkward and involves multiple allocations). `rotateCounterclockwise` is then added for symmetry and is a synonym for `perpendicularTo`.