Refactor curve evaluation functions
Add ability to conveniently evaluate curves at a list of parameter values, to avoid having to use List.map. This doesn't save a huge amount of typing, but it's a common use case and I think that

    Arc2d.pointsOn arc parameterValues

is significantly more readable than

    List.map (Arc2d.pointOn arc) parameterValues

This required renaming 'evaluate' to 'sample' since it was not clear how to pluralize 'evaluate'!